+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/domain/model/AgendaItem.kt
@@ -3,7 +3,7 @@ package com.a23pablooc.proxectofct.domain.model
 import java.util.Date
 
 data class AgendaItem(
-    var pkAgenda: Int,
+    var pkAgenda: Long,
     var descripcion: String,
     var fecha: Date
 )
