+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/domain/model/HistorialItem.kt
@@ -1,7 +1,7 @@
 package com.a23pablooc.proxectofct.domain.model
 
 data class HistorialItem(
-    val pkHistorial: Int,
+    val pkHistorial: Long,
     val fkMedicamento: MedicamentoItem,
     val fechaInicio: String,
     val fechaFin: String
