+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/domain/model/NotificacionItem.kt
@@ -3,7 +3,7 @@ package com.a23pablooc.proxectofct.domain.model
 import java.util.Date
 
 data class NotificacionItem(
-    var pkNotificacion: Int,
+    var pkNotificacion: Long,
     var fkMedicamento: MedicamentoItem,
     var fecha: Date,
     var hora: Date
