+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/domain/model/UsuarioItem.kt
@@ -1,6 +1,6 @@
 package com.a23pablooc.proxectofct.domain.model
 
 data class UsuarioItem(
-    val pkUsuario: Int,
+    val pkUsuario: Long,
     val nombre: String
 )
